Refactor assets layout
 * Move src/fonts/* into src/assets/fonts/* because they are assets.
 * Move authors' photos from src/assets/* to src/assets/photos/* to
   avoid mess and confusion.
 * Move src/assets/favicon.ico to src/assets/icons/*.
