+++ src/components/Header.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { NavLink } from 'react-router-dom';
 
 import '../styles/Header.styl';
-import Logo from '../assets/xsnippet.svg';
+import Logo from '../assets/icons/xsnippet.svg';
 
 const Header = () => (
   <header className="header" key="header">
