+++ src/styles/common/icons.styl
@@ -1,6 +1,6 @@
 @font-face
   font-family: "IconFont"
-  src: url("../fonts/icon-font.woff") format("woff")
+  src: url("../assets/fonts/icon-font.woff") format("woff")
   font-weight: normal
   font-style: normal
